Add support for signing a message without the coin header
XSN allows coin staking by creating a TPoS contract. This contract is included
into a transaction as an extra OP_RETURN output, it requires a signature from a
collateral UTXO (or OutPoint), this signature requires that the header is not
present (like Bitcoin Signed Message:).

The btchip_apdu_sign_message.c has been updated to add the new function for
signing a message without prepending a header.

Signed-off-by: Jonathan <jonathan@wiringbits.net>